+++ Rise Media Player Dev/ViewModels/SongViewModel.cs
@@ -9,6 +9,7 @@ using Rise.Common.Interfaces;
 using Rise.Data.ViewModels;
 using Rise.Models;
 using System;
+using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
@@ -335,28 +336,25 @@ namespace Rise.App.ViewModels
         /// <summary>
         /// Deletes item data from the backend.
         /// </summary>
-        public async Task DeleteAsync()
+        public async Task DeleteAsync(bool queue = false)
         {
-            App.MViewModel.Songs.Remove(this);
+            _ = App.MViewModel.Songs.Remove(this);
 
-            await NewRepository.Repository.DeleteAsync(Model);
+            if (queue)
+                NewRepository.Repository.QueueRemove(Model);
+            else
+                await NewRepository.Repository.DeleteAsync(Model);
 
-            AlbumViewModel album = App.MViewModel.Albums.
-                FirstOrDefault(a => a.Model.Title == Model.Album &&
-                           a.Model.Artist == Model.AlbumArtist);
+            IEnumerable<AlbumViewModel> albums = App.MViewModel.Albums.Where(a => a.Model.Title == Model.Album 
+                                    && a.Model.Artist == Model.AlbumArtist);
 
-            if (album != null)
-            {
-                await album.CheckAvailabilityAsync();
-            }
+            foreach (var album in albums)
+                await album.CheckAvailabilityAsync(queue);
 
-            ArtistViewModel artist = App.MViewModel.Artists.
-                FirstOrDefault(a => a.Model.Name == Model.Artist);
+            IEnumerable<ArtistViewModel> artists = App.MViewModel.Artists.Where(a => a.Model.Name == Model.Artist);
 
-            if (artist != null)
-            {
-                await artist.CheckAvailabilityAsync();
-            }
+            foreach (var artist in artists)
+                await artist.CheckAvailabilityAsync(queue);
         }
         #endregion
